feat: refactor, structure code, make DAGNode funcs inside the same folder, make tests pass again
