+++ package.json
@@ -39,6 +39,7 @@
   },
   "dependencies": {
     "async": "^2.1.2",
+    "buffer-loader": "0.0.1",
     "cids": "^0.2.0",
     "ipfs-block": "^0.5.0",
     "is-ipfs": "^0.2.1",
