+++ src/dag-node/index.js
@@ -16,7 +16,9 @@ class DAGNode {
     this._links = links || []
     this._serialized = serialized
     this._multihash = multihash
+
     this._size = this.links.reduce((sum, l) => sum + l.size, this.serialized.length)
+
     this._json = {
       data: this.data,
       links: this.links.map((l) => l.json),
@@ -39,7 +41,7 @@ class DAGNode {
   }
 
   set data (data) {
-    throw new Error(`Can't set property: 'data' is immutable`)
+    throw new Error("Can't set property: 'data' is immutable")
   }
 
   get links () {
@@ -47,7 +49,7 @@ class DAGNode {
   }
 
   set links (links) {
-    throw new Error(`Can't set property: 'links' is immutable`)
+    throw new Error("Can't set property: 'links' is immutable")
   }
 
   get serialized () {
@@ -55,7 +57,7 @@ class DAGNode {
   }
 
   set serialized (serialized) {
-    throw new Error(`Can't set property: 'serialized' is immutable`)
+    throw new Error("Can't set property: 'serialized' is immutable")
   }
 
   get size () {
@@ -63,7 +65,7 @@ class DAGNode {
   }
 
   set size (size) {
-    throw new Error(`Can't set property: 'size' is immutable`)
+    throw new Error("Can't set property: 'size' is immutable")
   }
 
   get multihash () {
@@ -71,8 +73,12 @@ class DAGNode {
   }
 
   set multihash (multihash) {
-    throw new Error(`Can't set property: 'multihash' is immutable`)
+    throw new Error("Can't set property: 'multihash' is immutable")
   }
 }
 
 exports = module.exports = DAGNode
+exports.create = require('./create')
+exports.clone = require('./clone')
+exports.addLink = require('./addLink')
+exports.rmLink = require('./rmLink')
